% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_scores.R
\name{aggregate_scores}
\alias{aggregate_scores}
\title{Aggregate deprivation scores}
\usage{
aggregate_scores(
  d,
  domain = "Index of Multiple Deprivation (IMD)",
  score_suffix = " Score",
  rank_suffix = " Rank (where 1 is most deprived)",
  decile_suffix = " Decile (where 1 is most deprived 10\% of LSOAs)",
  higher_level_geography,
  population_col = "No. people"
)
}
\arguments{
\item{d}{Dataframe containing deciles, ranks, scores and population 
estimates}

\item{domain}{Which domain to calculate population-weighted scores for}

\item{score_suffix}{The wording of the variable containing scores}

\item{rank_suffix}{The wording of the variable containing ranks}

\item{decile_suffix}{The wording of the variable containing deciles}

\item{higher_level_geography}{Name of the column to use for higher-level 
aggregation (e.g. "LAD19CD")}

\item{population_col}{Name of the column containing population estimates}
}
\value{
A tibble containing population-weighted average scores, 
    proportions of highly deprived areas, and the extent for
    the higher-level geography.
}
\description{
Aggregate deprivation scores into higher-level geographies,
    calculating:
    - proportion of highly deprived areas
    - extent (proportion of the local population that live in the most
      deprived areas)
    - population-weighted average score
}
\examples{
\dontrun{
aggregate_scores(imd_england, higher_level_geography = "LAD19CD")
}
}
