% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_weighted_scores.R
\name{calculate_pop_weighted_score}
\alias{calculate_pop_weighted_score}
\title{Population-weighted scores}
\usage{
calculate_pop_weighted_score(data, var, higher_level_geography, population)
}
\arguments{
\item{data}{Data frame containing a variable to be aggregated, lower level
geography population estimates, and a higher level geographical
grouping variable}

\item{var}{Name of the variable in the data frame containing the variable to
be aggregated (e.g. rank) for the lower level geography}

\item{higher_level_geography}{Name of the variable in the data frame
containing the higher level geography names/codes}

\item{population}{Name of the variable in the data frame containing
the population estimates of the lower level geography}
}
\description{
Calculate population-weighted scores within small areas.
}
\examples{
\dontrun{
calculate_pop_weighted_score(imd_england_lsoa, IMD_score, pop_count)
}
}
