% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census11_deprivation_england_wales_msoa11}
\alias{census11_deprivation_england_wales_msoa11}
\title{Census "Households by deprivation" data for MSOAs in England and Wales}
\format{
A data frame of class "tbl" with 36,005 rows and 4 variables:
\describe{
  \item{msoa11_code}{MSOA 2011 code}
  \item{households_number_deprivation_dimensions}{The number of dimensions of deprivation}
  \item{count}{Number of households}
  \item{proportion}{Proportion of households}
  ...
}
}
\source{
\url{https://www.nomisweb.co.uk/}
}
\usage{
census11_deprivation_england_wales_msoa11
}
\description{
This dataset provides Census 2011 estimates that classify households in
England and Wales by four dimensions of deprivation: Employment, education,
health and disability, and household overcrowding.
}
\keyword{datasets}
