% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census11_deprivation_northern_ireland_lgd14}
\alias{census11_deprivation_northern_ireland_lgd14}
\title{Census 2011 "Households by deprivation" data for Local Government Districtrs in Northern Ireland}
\format{
A data frame of class "tbl":
\describe{
  \item{lgd14_code}{Local Government District code}
  \item{households_number_deprivation_dimensions}{The number of dimensions of deprivation}
  \item{count}{Number of households}
  \item{proportion}{Proportion of households}
  ...
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
census11_deprivation_northern_ireland_lgd14
}
\description{
This dataset provides Census 2011 estimates that classify households in
Northern Ireland by four dimensions of deprivation: Employment, education,
health and disability, and household overcrowding.
}
\keyword{datasets}
