% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census21_deprivation_england_wales_lsoa21}
\alias{census21_deprivation_england_wales_lsoa21}
\title{Census "Households by deprivation" data for LSOAs in England and Wales}
\format{
A data frame of class "tbl" with 178,360 rows and 4 variables:
\describe{
  \item{lsoa21_code}{LSOA 2021 code}
  \item{households_number_deprivation_dimensions}{The number of dimensions of deprivation}
  \item{count}{Count of households}
  \item{percent}{Percentage of households}
  ...
}
}
\source{
\url{https://www.nomisweb.co.uk/}
}
\usage{
census21_deprivation_england_wales_lsoa21
}
\description{
This dataset provides Census 2021 estimates that classify households in
England and Wales by four dimensions of deprivation: Employment, education,
health and disability, and household overcrowding.
}
\keyword{datasets}
