% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census21_deprivation_northern_ireland_sdz21}
\alias{census21_deprivation_northern_ireland_sdz21}
\title{Census 2021 "Households by deprivation" data for Super Data Zones in Northern Ireland}
\format{
A data frame of class "tbl":
\describe{
  \item{sdz21_code}{Super Data Zone code}
  \item{households_number_deprivation_dimensions}{The number of dimensions of deprivation}
  \item{count}{Number of households}
  \item{proportion}{Proportion of households}
  ...
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
census21_deprivation_northern_ireland_sdz21
}
\description{
This dataset provides Census 2021 estimates that classify households in
Northern Ireland by four dimensions of deprivation: Employment, education,
health and disability, and household overcrowding.
}
\keyword{datasets}
