% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cni2023_england_lsoa21}
\alias{cni2023_england_lsoa21}
\title{Community Needs Index and Left Behind Areas for LSOAs (2021 codes) in England}
\format{
A data frame of class "tbl" with 33,755 rows and 11 variables:
\describe{
  \item{lsoa21_code}{LSOA code}
  \item{lsoa21_name}{LSOA name}
  \item{Community Needs Index 2023 Score}{Overall Community Needs Index score (higher score = higher needs)}
  \item{Community Needs Index 2023 Rank}{Overall Community Needs Index rank (1 = highest needs)}
  \item{Civic Assets Domain Score}{Score for the Civic Assets domain (higher score = higher needs)}
  \item{Civic Assets Domain Rank}{Rank for the Civic Assets domain (1 = highest needs)}
  \item{Connectedness Domain Score}{Score for the Connectedness domain (higher score = higher needs)}
  \item{Connectedness Domain Rank}{Rank for the Connectedness domain (1 = highest needs)}
  \item{Active and Engaged Community Domain Score}{Score for the Engaged Community domain (higher score = higher needs)}
  \item{Engaged and Engaged Community Domain Rank}{Rank for the Engaged Community domain (1 = highest needs)}
  \item{Left Behind Area?}{TRUE if this Ward is a Left Behind Area}
  ...
}
}
\source{
\url{https://ocsi.uk/}
}
\usage{
cni2023_england_lsoa21
}
\description{
A data set containing the Community Needs Index scores and ranks and its
three domains. Also highlights which areas are categorised as 'Left Behind'.
}
\keyword{datasets}
