% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cni_northern_ireland_soa11}
\alias{cni_northern_ireland_soa11}
\title{Community Needs Index and Left Behind Areas for Super Output Areas (2011
codes) in Northern Ireland}
\format{
A data frame of class "tbl" with 890 rows and 13 variables:
\describe{
  \item{soa11_code}{Super Output Area code}
  \item{soa11_name}{Super Output Area name}
  \item{lgd14_name}{Local Government District name}
  \item{lgd14_code}{Local Government District code}
  \item{Civic Assets score}{Score for the Civic Assets domain (higher score = higher needs)}
  \item{Civic Assets rank}{Rank for the Civic Assets domain (1 = highest needs)}
  \item{Connectedness score}{Score for the Connectedness domain (higher score = higher needs)}
  \item{Connectedness rank}{Rank for the Connectedness domain (1 = highest needs)}
  \item{Engaged Community score}{Score for the Engaged Community domain (higher score = higher needs)}
  \item{Engaged Community rank}{Rank for the Engaged Community domain (1 = highest needs)}
  \item{Community Needs Index score}{Overall Community Needs Index score (higher score = higher needs)}
  \item{Community Needs Index rank}{Overall Community Needs Index rank (1 = highest needs)}
  \item{Left Behind Area?}{TRUE if this Ward is a Left Behind Area}
  ...
}
}
\source{
\url{https://www.osci.uk/}
}
\usage{
cni_northern_ireland_soa11
}
\description{
A data set containing the Community Needs Index scores and ranks and its
three domains. Also highlights which areas are categorised as 'Left Behind'.
}
\keyword{datasets}
