% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cni_scotland_iz11}
\alias{cni_scotland_iz11}
\title{Community Needs Index and Left Behind Areas for Intermediate Zones (2011
codes) in Scotland}
\format{
A data frame of class "tbl" with 1,279 rows and 11 variables:
\describe{
  \item{iz11_code}{Intermediate Zone code}
  \item{iz11_name}{Intermediate Zone name}
  \item{Civic Assets score}{Score for the Civic Assets domain (higher score = higher needs)}
  \item{Civic Assets rank}{Rank for the Civic Assets domain (1 = highest needs)}
  \item{Connectedness score}{Score for the Connectedness domain (higher score = higher needs)}
  \item{Connectedness rank}{Rank for the Connectedness domain (1 = highest needs)}
  \item{Engaged Community score}{Score for the Engaged Community domain (higher score = higher needs)}
  \item{Engaged Community rank}{Rank for the Engaged Community domain (1 = highest needs)}
  \item{Community Needs Index score}{Overall Community Needs Index score (higher score = higher needs)}
  \item{Community Needs Index rank}{Overall Community Needs Index rank (1 = highest needs)}
  \item{Left Behind Area?}{TRUE if this Ward is a Left Behind Area}
  ...
}
}
\source{
\url{https://ocsi.uk/}
}
\usage{
cni_scotland_iz11
}
\description{
A data set containing the Community Needs Index scores and ranks and its
three domains. Also highlights which areas are categorised as 'Left Behind'.
}
\keyword{datasets}
