% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cni_wales_msoa11}
\alias{cni_wales_msoa11}
\title{Community Needs Index and Left Behind Areas for MSOAs (2011 codes) in Wales}
\format{
A data frame of class "tbl" with 410 rows and 12 variables:
\describe{
  \item{msoa11_code}{MSOA code}
  \item{msoa11_name}{MSOA name}
  \item{lad21_name}{Local Authority name}
  \item{Civic Assets score}{Score for the Civic Assets domain (higher score = higher needs)}
  \item{Civic Assets rank}{Rank for the Civic Assets domain (1 = highest needs)}
  \item{Connectedness score}{Score for the Connectedness domain (higher score = higher needs)}
  \item{Connectedness rank}{Rank for the Connectedness domain (1 = highest needs)}
  \item{Engaged Community score}{Score for the Engaged Community domain (higher score = higher needs)}
  \item{Engaged Community rank}{Rank for the Engaged Community domain (1 = highest needs)}
  \item{Community Needs Index score}{Overall Community Needs Index score (higher score = higher needs)}
  \item{Community Needs Index rank}{Overall Community Needs Index rank (1 = highest needs)}
  \item{Left Behind Area?}{TRUE if this Ward is a Left Behind Area}
  ...
}
}
\source{
\url{https://www.osci.uk/}
}
\usage{
cni_wales_msoa11
}
\description{
A data set containing the Community Needs Index scores and ranks and its
three domains. Also highlights which areas are categorised as 'Left Behind'.
}
\keyword{datasets}
