% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2000_england_ward98}
\alias{imd2000_england_ward98}
\title{Index of Multiple Deprivation (2000) for Wards in England}
\format{
A data frame of class "tbl" with 8,414 rows and 21 variables:
\describe{
  \item{Ward}{Ward code}
  \item{Ward Name}{Ward name}
  \item{LA}{Local Authority code}
  \item{LA Name}{LA name}
  \item{DETR LA Code}{DETR LA code}
  \item{Index of Multiple Deprivation Score}{IMD score}
  \item{Rank of Index of Multiple Deprivation}{IMD rank}
  \item{Income Domain Score}{Income domain score}
  \item{Rank of Income Domain}{Income domain rank}
  \item{Employment Domain Score}{Employment domain score}
  \item{Rank of Employment Domain}{Employment domain rank}
  \item{Health Domain Score}{Health domain score}
  \item{Rank of Health Domain}{Health domain rank}
  \item{Education Domain Score}{Education domain score}
  \item{Rank of Education Domain}{Education domain rank}
  \item{Housing Domain Score}{Housing domain score}
  \item{Rank of Housing Domain}{Housing domain rank}
  \item{Access Domain Score}{Access domain score}
  \item{Rank of Access Domain}{Access domain rank}
  \item{Child Poverty Index Score}{Child Poverty Index score}
  \item{Rank of Child Poverty Index}{Child Poverty Index rank}
  \item{IMD_decile}{IMD decile}
  \item{Income_decile}{Income domain decile}
  \item{Employment_decile}{Employment domain decile}
  \item{Health_decile}{Health domain decile}
  \item{Education_decile}{Education domain decile}
  \item{Housing_decile}{Housing domain decile}
  \item{Access_decile}{Access to services domain decile}
  \item{Child_Poverty_decile}{Child Poverty Index decile}
  ...
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd2000_england_ward98
}
\description{
A data set containing IMD scores, ranks and deciles for Wards in England.
}
\keyword{datasets}
