% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2010_england_msoa11}
\alias{imd2010_england_msoa11}
\title{Index of Multiple Deprivation (2010) for MSOA in England}
\format{
A data frame of class "tbl" with 6,791 rows and 4 variables:
\describe{
  \item{msoa11_code}{MSOA code}
  \item{Score}{Population-weighted average IMD score}
  \item{Proportion}{Proportion of the nation's most deprived small areas in each LA}
  \item{Extent}{Extent of the LAs population living in the most deprived areas}
  ...
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd2010_england_msoa11
}
\description{
A data set containing IMD scores, proportions and extents for Local MSOAs
in England.
}
\keyword{datasets}
