% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2010_lsoa01_england}
\alias{imd2010_lsoa01_england}
\title{Index of Multiple Deprivation (2010) for LSOAs in England}
\format{
A data frame of class "tbl" with 32,482 rows and 7 variables:
\describe{
  \item{lsoa01_code}{LSOA (2001) code}
  \item{lad09_code}{Local Authority (2009) code}
  \item{lad09_name}{Local Authority name}
  \item{IMD_score}{Index of multiple deprivation score}
  \item{IMD_rank}{Index of multiple deprivation rank}
  \item{IMD_decile}{Index of multiple deprivation decile}
  ...
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd2010_lsoa01_england
}
\description{
A data set containing IMD ranks and deciles for LSOAs in England.
}
\keyword{datasets}
