% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2010_northern_ireland_soa01}
\alias{imd2010_northern_ireland_soa01}
\title{Index of Multiple Deprivation (2010) for Super Output Areas in Northern Ireland}
\format{
A data frame of class "tbl" with 890 rows and 17 variables:
\describe{
  \item{soa01_code}{Super Output Area code}
  \item{IMD_rank}{IMD rank}
  \item{Income_rank}{Income domain rank}
  \item{Employment_rank}{Employment domain rank}
  \item{Health_rank}{Health domain rank}
  \item{Education_rank}{Education domain rank}
  \item{Access_rank}{Access to services domain rank}
  \item{Environment_rank}{Environment domain rank}
  \item{Crime_rank}{Crime domain rank}
  \item{IMD_decile}{IMD decile}
  \item{Income_decile}{Income domain decile}
  \item{Employment_decile}{Employment domain decile}
  \item{Health_decile}{Health and Disability domain decile}
  \item{Education_decile}{Education domain decile}
  \item{Access_decile}{Access to services domain decile}
  \item{Environment_decile}{Environment domain decile}
  \item{Crime_decile}{Crime domain decile}
  ...
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
imd2010_northern_ireland_soa01
}
\description{
A data set containing IMD ranks and deciles for SOAs in Northern Ireland.
}
\keyword{datasets}
