% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2011_wales_lsoa01}
\alias{imd2011_wales_lsoa01}
\title{Index of Multiple Deprivation (2011) for LSOAs in Wales}
\format{
A data frame of class "tbl" with 1,896 rows and 19 variables:
\describe{
  \item{lsoa01_code}{LSOA code}
  \item{IMD_rank}{IMD rank}
  \item{Income_rank}{Income domain rank}
  \item{Employment_rank}{Employment domain rank}
  \item{Education_rank}{Education domain rank}
  \item{Health_rank}{Health domain rank}
  \item{Crime_rank}{Crime domain rank}
  \item{Housing_rank}{Housing domain rank}
  \item{Access_rank}{Access to services domain rank}
  \item{Environment_rank}{Environment domain rank}
  \item{IMD_decile}{IMD decile}
  \item{Income_decile}{Income domain decile}
  \item{Employment_decile}{Employment domain decile}
  \item{Education_decile}{Education domain decile}
  \item{Health_decile}{Health and Disability domain decile}
  \item{Crime_decile}{Crime domain decile}
  \item{Housing_decile}{Housing domain decile}
  \item{Access_decile}{Access to services domain decile}
  \item{Environment_decile}{Environment domain decile}
  ...
}
}
\source{
\url{https://www.gov.wales/}
}
\usage{
imd2011_wales_lsoa01
}
\description{
A data set containing IMD ranks and deciles for Lower Layer Super
Output Areas (LSOAs) in Wales.
}
\keyword{datasets}
