% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2012_scotland_dz01}
\alias{imd2012_scotland_dz01}
\title{Index of Multiple Deprivation (2012) for Data Zones in Scotland}
\format{
A data frame of class "tbl" with 6,505 rows and 17 variables:
\describe{
  \item{dz01_code}{Data Zone code}
  \item{IMD_rank}{IMD rank}
  \item{Income_rank}{Income domain rank}
  \item{Employment_rank}{Employment domain rank}
  \item{Health_rank}{Health domain rank}
  \item{Education_rank}{Education domain rank}
  \item{Housing_rank}{Housing domain rank}
  \item{Access_rank}{Access to services domain rank}
  \item{Crime_rank}{Crime domain rank}
  \item{IMD_decile}{IMD decile}
  \item{Income_decile}{Income domain decile}
  \item{Employment_decile}{Employment domain decile}
  \item{Health_decile}{Health and Disability domain decile}
  \item{Education_decile}{Education domain decile}
  \item{Access_decile}{Access to services domain decile}
  \item{Housing_decile}{Housing domain decile}
  \item{Crime_decile}{Crime domain decile}
  ...
}
}
\source{
\url{https://www.gov.scot/}
}
\usage{
imd2012_scotland_dz01
}
\description{
A data set containing IMD ranks and deciles for Data Zones in Scotland
}
\keyword{datasets}
