% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2015_lad21_england}
\alias{imd2015_lad21_england}
\title{Index of Multiple Deprivation (2015) for Local Authorities in England}
\format{
A data frame of class "tbl" with 309 rows and 3 variables:
\describe{
  \item{ltla21_code}{Local Authority (2021) code}
  \item{Score}{Population-weighted deprivation score}
  \item{Proportion}{Proportion of the nation's most deprived small areas in each LA}
  \item{Extent}{Extent of the LAs population living in the most deprived areas}
  ...
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd2015_lad21_england
}
\description{
A data set containing IMD proportions and extents for Local Authorites
(2021 codes) in England.
}
\keyword{datasets}
