% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2015_lsoa11_england}
\alias{imd2015_lsoa11_england}
\title{Index of Multiple Deprivation (2015) for LSOAs in England}
\format{
A data frame of class "tbl" with 32,844 rows and 6 variables:
\describe{
  \item{lsoa11_code}{LSOA (2011) code}
  \item{lsoa11_name}{LSOA name}
  \item{lad13_code}{Local Authority (2013) code}
  \item{lad13_name}{Local Authority name}
  \item{IMD_rank}{Index of multiple deprivation rank}
  \item{IMD_decile}{Index of multiple deprivation decile}
  ...
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd2015_lsoa11_england
}
\description{
A data set containing IMD ranks and deciles for LSOAs in England.
}
\keyword{datasets}
