% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2015_msoa11_england}
\alias{imd2015_msoa11_england}
\title{Index of Multiple Deprivation (2015) for MSOAs in England}
\format{
A data frame of class "tbl" with 6,791 rows and 3 variables:
\describe{
  \item{msoa111_code}{MSOA code}
  \item{Proportion}{Proportion of the nation's most deprived small areas in each LA}
  \item{Extent}{Extent of the LAs population living in the most deprived areas}
  ...
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd2015_msoa11_england
}
\description{
A data set containing IMD proportions and extents for MSOAs in England.
}
\keyword{datasets}
