% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2016_scotland_dz11_indicators}
\alias{imd2016_scotland_dz11_indicators}
\title{Underlying indicators for Scottish Index of Multiple Deprivation 2016}
\format{
A data frame of class "tbl" with 6,976 rows and 34 variables:
\describe{
  \item{dz11_code}{Data Zone code}
  \item{Total_population}{Total population}
  \item{Working_age_population_revised}{Working age population}
  \item{Income_rate}{Percentage of people who are income deprived}
  \item{Income_count}{Number of people who are income deprived}
  \item{Employment_rate}{Percentage of people who are employment deprived}
  \item{Employment_count}{Number of people who are employment deprived}
  \item{CIF}{Comparative Illness Factor: standardised ratio}
  \item{ALCOHOL}{Hospital stays related to alcohol misuse: standardised ratio}
  \item{DRUG}{Hospital stays related to drug misuse: standardised ratio}
  \item{SMR}{Standardised mortality ratio}
  \item{DEPRESS}{Proportion of population being prescribed drugs for anxiety, depression or psychosis}
  \item{LBWT}{Proportion of live singleton births of low birth weight}
  \item{EMERG}{Emergency stays in hospital: standardised ratio}
  \item{Attendance}{School pupil attendance}
  \item{Attainment}{Attainment of school leavers}
  \item{Noquals}{Working age people with no qualifications: standardised ratio}
  \item{NEET}{Proportion of people aged 16-19 not in full time education, employment or training}
  \item{HESA}{Proportion of 17-21 year olds entering in to full time higher education}
  \item{drive_petrol}{Average drive time to a petrol station in minutes}
  \item{drive_GP}{Average drive time to a GP surgery in minutes}
  \item{drive_PO}{Average drive time to a post office in minutes}
  \item{drive_primary}{Average drive time to a primary school in minutes}
  \item{drive_retail}{Average drive time to a retail centre in minutes}
  \item{drive_secondary}{Average drive time to a secondary school in minutes}
  \item{PT_GP}{Public transport travel time to a GP surgery in minutes}
  \item{PT_Post}{Public transport travel time to a post office in minutes}
  \item{PT_retail}{Public transport travel time to a retail centre in minutes}
  \item{crime_count}{Number of recorded crimes of violence, sexual offences, domestic housebreaking, vandalism, drugs offences, and common assault}
  \item{crime_rate}{Recorded crimes of violence, sexual offences, domestic housebreaking, vandalism, drugs offences, and common assault per 10,000 people}
  \item{overcrowded_count}{Number of people in households that are overcrowded}
  \item{nocentralheat_count}{Number of people in households without central heating}
  \item{overcrowded_rate}{Percentage of people in households that are overcrowded}
  \item{nocentralheat_rate}{Percentage of people in households without central heating}
  ...
}
}
\source{
\url{https://www.gov.scot/}
}
\usage{
imd2016_scotland_dz11_indicators
}
\description{
A data set containing proportions and extents for Council Areas (Local
Authorities) in Scotland. Note that IMD scores for Scotland are not
available.
}
\keyword{datasets}
