% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2017_northern_ireland_soa01_indicators}
\alias{imd2017_northern_ireland_soa01_indicators}
\title{Northern Ireland Multiple Deprivation Measure 2017 underlying indicators}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 890 rows and 81 columns.
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
imd2017_northern_ireland_soa01_indicators
}
\description{
Northern Ireland Multiple Deprivation Measure 2017 underlying indicators
}
\keyword{datasets}
