% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2019_england_lsoa11}
\alias{imd2019_england_lsoa11}
\title{Index of Multiple Deprivation for LSOAs (2011) in England}
\format{
A data frame of class "tbl" with 32,844 rows and 17 variables:
\describe{
  \item{lsoa11_code}{LSOA code}
  \item{IMD_decile}{IMD decile}
  \item{Income_decile}{Income domain decile}
  \item{Employment_decile}{Employment domain decile}
  \item{Education_decile}{Education domain decile}
  \item{Health_decile}{Health and Disability domain decile}
  \item{Crime_decile}{Crime domain decile}
  \item{Housing_and_Access_decile}{Housing and access to services domain decile}
  \item{Environment_decile}{Environment domain decile}
  \item{IMD_rank}{IMD rank}
  \item{Income_rank}{Income domain rank}
  \item{Employment_rank}{Employment domain rank}
  \item{Education_rank}{Education domain rank}
  \item{Health_rank}{Health domain rank}
  \item{Crime_rank}{Crime domain rank}
  \item{Housing_and_Access_rank}{Housing and access to services domain rank}
  \item{Environment_rank}{Environment domain rank}
  ...
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd2019_england_lsoa11
}
\description{
A data set containing IMD ranks and deciles for Lower Layer Super
Output Areas (LSOAs) in England.
}
\keyword{datasets}
