% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2019_england_lsoa11_indicators}
\alias{imd2019_england_lsoa11_indicators}
\title{Underlying indicators for Index of Multiple Deprivation for LSOAs (2011) in England}
\format{
A data frame of class "tbl" with 32,844 rows and 29 variables:
\describe{
  \item{lsoa11_code}{lsoa11_code}
  \item{income_domain_numerator}{Income Domain numerator}
  \item{income_deprivation_affecting_children_index_idaci_numerator}{Income Deprivation Affecting Children Index (IDACI) numerator}
  \item{income_deprivation_affecting_older_people_index_idaopi_numerator}{Income Deprivation Affecting Older People Index (IDAOPI) numerator}
  \item{employment_domain_numerator}{Employment Domain numerator}
  \item{staying_on_in_education_post_16_indicator}{Staying on in education post 16 indicator}
  \item{entry_to_higher_education_indicator}{Entry to higher education indicator}
  \item{adult_skills_and_english_language_proficiency_indicator}{Adult skills and English language proficiency indicator}
  \item{years_of_potential_life_lost_indicator}{Years of potential life lost indicator}
  \item{comparative_illness_and_disability_ratio_indicator}{Comparative illness and disability ratio indicator}
  \item{acute_morbidity_indicator}{Acute morbidity indicator}
  \item{mood_and_anxiety_disorders_indicator}{Mood and anxiety disorders indicator}
  \item{road_distance_to_a_post_office_indicator_km}{Road distance to a post office indicator (km)}
  \item{road_distance_to_a_primary_school_indicator_km}{Road distance to a primary school indicator (km)}
  \item{road_distance_to_general_store_or_supermarket_indicator_km}{Road distance to general store or supermarket indicator (km)}
  \item{road_distance_to_a_gp_surgery_indicator_km}{Road distance to a GP surgery indicator (km)}
  \item{household_overcrowding_indicator}{Household overcrowding indicator}
  \item{homelessness_indicator_rate_per_1000_households}{Homelessness indicator (rate per 1000 households)}
  \item{owner_occupation_affordability_component_of_housing_affordability_indicator}{Owner-occupation affordability (component of housing affordability indicator)}
  \item{private_rental_affordability_component_of_housing_affordability_indicator}{Private rental affordability (component of housing affordability indicator)}
  \item{housing_affordability_indicator}{Housing affordability indicator}
  \item{housing_in_poor_condition_indicator}{Housing in poor condition indicator}
  \item{houses_without_central_heating_indicator}{Houses without central heating indicator}
  \item{road_traffic_accidents_indicator}{Road traffic accidents indicator}
  \item{nitrogen_dioxide_component_of_air_quality_indicator}{Nitrogen dioxide (component of air quality indicator)}
  \item{benzene_component_of_air_quality_indicator}{Benzene (component of air quality indicator)}
  \item{sulphur_dioxide_component_of_air_quality_indicator}{Sulphur dioxide (component of air quality indicator)}
  \item{particulates_component_of_air_quality_indicator}{Particulates (component of air quality indicator)}
  \item{air_quality_indicator}{Air quality indicator}
  ...
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd2019_england_lsoa11_indicators
}
\description{
A data set containing IMD scores and deciles for Lower Layer Super
Output Areas (LSOAs) in England.
}
\keyword{datasets}
