% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2019_england_ltla19}
\alias{imd2019_england_ltla19}
\title{Index of Multiple Deprivation (2019) for Local Authorities (2019) in England}
\format{
A data frame of class "tbl" with 317 rows and 18 variables:
\describe{
  \item{ltla19_code}{LAD code}
  \item{Score}{Population-weighted average score}
  \item{Proportion}{Proportion of small areas in the nation's most deprived}
  \item{Extent}{Extent for overall IMD}
  \item{Income_Score}{Population-weighted average score for Income domain}
  \item{Income_Proportion}{Proportion of most deprived areas for Income}
  \item{Employment_Score}{Population-weighted average score for Employment
    domain}
  \item{Employment_Proportion}{Proportion of most deprived areas for
    Employment}
  \item{Education_Score}{Population-weighted average score for Education
    domain}
  \item{Education_Proportion}{Proportion of most deprived areas for
    Education}
  \item{Health_Score}{Population-weighted average score for Health domain}
  \item{Health_Proportion}{Proportion of most deprived areas for Health}
  \item{Crime_Score}{Population-weighted average score for Crime domain}
  \item{Crime_Proportion}{Proportion of most deprived areas for Crime}
  \item{Housing_and_Access_Score}{Population-weighted average score for
    Housing/Access domain}
  \item{Housing_and_Access_Proportion}{Proportion of most deprived areas
    for Housing/Access domain}
  \item{Environment_Score}{Population-weighted average score for Environment
    domain}
  \item{Environment_Proportion}{Proportion of most deprived areas for
    Environment}
  ...
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd2019_england_ltla19
}
\description{
A data set containing aggregated scores and proportions for Local
Authorities in England. Note that extents are not available.
}
\keyword{datasets}
