% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2020_scotland_dz11_indicators}
\alias{imd2020_scotland_dz11_indicators}
\title{Underlying indicators for Scottish Index of Multiple Deprivation 2020}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6976 rows and 35 columns.
}
\source{
\url{https://www.gov.scot/}
}
\usage{
imd2020_scotland_dz11_indicators
}
\description{
Underlying indicators for Scottish Index of Multiple Deprivation 2020
}
\keyword{datasets}
