% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2020_scotland_iz11}
\alias{imd2020_scotland_iz11}
\title{Index of Multiple Deprivation for IZs in Scotland}
\format{
A data frame of class "tbl" with 1,279 rows and 17 variables:
\describe{
  \item{iz11_code}{Intermediate Zone code}
  \item{Proportion}{Proportion of small areas in the nation's most deprived}
  \item{Extent}{Extent for overall IMD}
  \item{Income_Proportion}{Proportion of most deprived areas for Income}
  \item{Income_Extent}{Extent for Income}
  \item{Employment_Proportion}{Proportion of most deprived areas for
    Employment}
  \item{Employment_Extent}{Extent for Employment}
  \item{Education_Proportion}{Proportion of most deprived areas for
    Education}
  \item{Education_Extent}{Extent for Education}
  \item{Health_Proportion}{Proportion of most deprived areas for Health}
  \item{Health_Extent}{Extent for Health}
  \item{Crime_Proportion}{Proportion of most deprived areas for Crime}
  \item{Crime_Extent}{Extent for Crime}
  \item{Housing_Proportion}{Proportion of most deprived areas
    for Housing domain}
  \item{Housing_Extent}{Extent for Housing}
  \item{Access_Proportion}{Proportion of most deprived areas
    for Access domain}
  \item{Access_Extent}{Extent for Access}
  ...
}
}
\source{
\url{https://www.gov.scot/}
}
\usage{
imd2020_scotland_iz11
}
\description{
A data set containing proportions and extents for Intermediate Zones
(IZs) in Scotland. Note that IMD scores for Scotland are not available.
}
\keyword{datasets}
