% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2025_england_lrf}
\alias{imd2025_england_lrf}
\title{Index of Multiple Deprivation (2025) for LRFs (2024) in England}
\format{
A data frame of class "tbl":
\describe{
  \item{lrf24_code}{LRF code}
  \item{lrf24_name}{LRF name}
  \item{IMD - Average rank}{IMD - Average rank}
  \item{IMD - Rank of average rank}{IMD - Rank of average rank}
  \item{IMD - Average score}{IMD - Average score}
  \item{IMD - Rank of average score}{IMD - Rank of average score}
  \item{IMD - Proportion of LSOAs in most deprived 10percent nationally}{IMD - Proportion of LSOAs in most deprived 10 percent nationally}
  \item{IMD - Rank of proportion of LSOAs in most deprived 10percent nationally}{IMD - Rank of proportion of LSOAs in most deprived 10 percent nationally}
  \item{IMD 2025 - Extent}{IMD 2025 - Extent}
  \item{IMD 2025 - Rank of extent}{IMD 2025 - Rank of extent}
  \item{IMD 2025 - Local concentration}{IMD 2025 - Local concentration}
  \item{IMD 2025 - Rank of local concentration}{IMD 2025 - Rank of local concentration}
  \item{Income - Average rank}{Income - Average rank}
  \item{Income - Rank of average rank}{Income - Rank of average rank}
  \item{Income - Average score}{Income - Average score}
  \item{Income - Rank of average score}{Income - Rank of average score}
  \item{Income - Proportion of LSOAs in most deprived 10percent nationally}{Income - Proportion of LSOAs in most deprived 10 percent nationally}
  \item{Income - Rank of proportion of LSOAs in most deprived 10percent nationally}{Income - Rank of proportion of LSOAs in most deprived 10 percent nationally}
  \item{Income scale}{Income scale}
  \item{Rank of Income Scale}{Rank of Income Scale}
  \item{Employment - Average rank}{Employment - Average rank}
  \item{Employment - Rank of average rank}{Employment - Rank of average rank}
  \item{Employment - Average score}{Employment - Average score}
  \item{Employment - Rank of average score}{Employment - Rank of average score}
  \item{Employment - Proportion of LSOAs in most deprived 10percent nationally}{Employment - Proportion of LSOAs in most deprived 10 percent nationally}
  \item{Employment - Rank of proportion of LSOAs in most deprived 10percent nationally}{Employment - Rank of proportion of LSOAs in most deprived 10 percent nationally}
  \item{Employment scale}{Employment scale}
  \item{Rank of Employment Scale}{Rank of Employment Scale}
  \item{Education, Skills and Training - Average rank}{Education, Skills and Training - Average rank}
  \item{Education, Skills and Training - Rank of average rank}{Education, Skills and Training - Rank of average rank}
  \item{Education, Skills and Training - Average score}{Education, Skills and Training - Average score}
  \item{Education, Skills and Training - Rank of average score}{Education, Skills and Training - Rank of average score}
  \item{Education, Skills and Training - Proportion of LSOAs in most deprived 10percent nationally}{Education, Skills and Training - Proportion of LSOAs in most deprived 10 percent nationally}
  \item{Education, Skills and Training - Rank of proportion of LSOAs in most deprived 10percent nationally}{Education, Skills and Training - Rank of proportion of LSOAs in most deprived 10 percent nationally}
  \item{Health Deprivation and Disability - Average rank}{Health Deprivation and Disability - Average rank}
  \item{Health Deprivation and Disability - Rank of average rank}{Health Deprivation and Disability - Rank of average rank}
  \item{Health Deprivation and Disability - Average score}{Health Deprivation and Disability - Average score}
  \item{Health Deprivation and Disability - Rank of average score}{Health Deprivation and Disability - Rank of average score}
  \item{Health Deprivation and Disability - Proportion of LSOAs in most deprived 10percent nationally}{Health Deprivation and Disability - Proportion of LSOAs in most deprived 10 percent nationally}
  \item{Health Deprivation and Disability - Rank of proportion of LSOAs in most deprived 10percent nationally}{Health Deprivation and Disability - Rank of proportion of LSOAs in most deprived 10 percent nationally}
  \item{Crime - Average rank}{Crime - Average rank}
  \item{Crime - Rank of average rank}{Crime - Rank of average rank}
  \item{Crime - Average score}{Crime - Average score}
  \item{Crime - Rank of average score}{Crime - Rank of average score}
  \item{Crime - Proportion of LSOAs in most deprived 10percent nationally}{Crime - Proportion of LSOAs in most deprived 10 percent nationally}
  \item{Crime - Rank of proportion of LSOAs in most deprived 10percent nationally}{Crime - Rank of proportion of LSOAs in most deprived 10 percent nationally}
  \item{Barriers to Housing and Services - Average rank}{Barriers to Housing and Services - Average rank}
  \item{Barriers to Housing and Services - Rank of average rank}{Barriers to Housing and Services - Rank of average rank}
  \item{Barriers to Housing and Services - Average score}{Barriers to Housing and Services - Average score}
  \item{Barriers to Housing and Services - Rank of average score}{Barriers to Housing and Services - Rank of average score}
  \item{Barriers to Housing and Services - Proportion of LSOAs in most deprived 10percent nationally}{Barriers to Housing and Services - Proportion of LSOAs in most deprived 10 percent nationally}
  \item{Barriers to Housing and Services - Rank of proportion of LSOAs in most deprived 10percent nationally}{Barriers to Housing and Services - Rank of proportion of LSOAs in most deprived 10 percent nationally}
  \item{Living Environment - Average rank}{Living Environment - Average rank}
  \item{Living Environment - Rank of average rank}{Living Environment - Rank of average rank}
  \item{Living Environment - Average score}{Living Environment - Average score}
  \item{Living Environment - Rank of average score}{Living Environment - Rank of average score}
  \item{Living Environment - Proportion of LSOAs in most deprived 10percent nationally}{Living Environment - Proportion of LSOAs in most deprived 10 percent nationally}
  \item{Living Environment - Rank of proportion of LSOAs in most deprived 10percent nationally}{Living Environment - Rank of proportion of LSOAs in most deprived 10 percent nationally}
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd2025_england_lrf
}
\description{
A dataset containing IMD scores, ranks, and related measures for Local Resilience Forums in England.
}
\keyword{datasets}
