% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2025_england_lsoa21_indicators}
\alias{imd2025_england_lsoa21_indicators}
\title{Index of Multiple Deprivation (2025) Indicators for LSOAs (2021) in England}
\format{
A data frame of class "tbl":
\describe{
  \item{lsoa21_code}{LSOA 2021 code}
  \item{Income Domain numerator}{Income Domain numerator}
  \item{Income Deprivation Affecting Children Index (IDACI) numerator}{Income Deprivation Affecting Children Index (IDACI) numerator}
  \item{Income Deprivation Affecting Older People Index (IDAOPI) numerator}{Income Deprivation Affecting Older People Index (IDAOPI) numerator}
  \item{Employment Domain numerator}{Employment Domain numerator}
  \item{Adult skills and English language proficiency indicator}{Adult skills and English language proficiency indicator}
  \item{Entry to higher education indicator}{Entry to higher education indicator}
  \item{Acute morbidity indicator}{Acute morbidity indicator}
  \item{Comparative illness and disability ratio indicator}{Comparative illness and disability ratio indicator}
  \item{Mental health indicator}{Mental health indicator}
  \item{Years of potential life lost indicator}{Years of potential life lost indicator}
  \item{Violence with injury, rate per 1,000 'at-risk population'}{Violence with injury, rate per 1,000 at-risk population}
  \item{Violence without injury, rate per 1,000 'at-risk population'}{Violence without injury, rate per 1,000 at-risk population}
  \item{Stalking and harassment, rate per 1,000 'at-risk population'}{Stalking and harassment, rate per 1,000 at-risk population}
  \item{Burglary, rate per 1,000 'at-risk properties'}{Burglary, rate per 1,000 at-risk properties}
  \item{Theft, rate per 1,000 'at-risk population'}{Theft, rate per 1,000 at-risk population}
  \item{Criminal damage, rate per 1,000 'at-risk population'}{Criminal damage, rate per 1,000 at-risk population}
  \item{Public order and Possession of weapons, rate per 1,000 'at-risk population'}{Public order and Possession of weapons, rate per 1,000 at-risk population}
  \item{Anti-Social Behaviour, rate per 1,000 'at-risk population'}{Anti-Social Behaviour, rate per 1,000 at-risk population}
  \item{Connectivity Score}{Connectivity Score}
  \item{Owner-occupation affordability (component of housing affordability indicator)}{Owner-occupation affordability (component of housing affordability indicator)}
  \item{Private rental affordability (component of housing affordability indicator)}{Private rental affordability (component of housing affordability indicator)}
  \item{Housing affordability indicator}{Housing affordability indicator}
  \item{Digital Connectivity (Broadband download and upload speeds)}{Digital Connectivity (Broadband download and upload speeds)}
  \item{Core homelessness rate (percent of households)}{Core homelessness rate (percent of households)}
  \item{Patient-to-GP Ratio}{Patient-to-GP Ratio}
  \item{Homelessness indicator (rate per 1,000 households)}{Homelessness indicator (rate per 1,000 households)}
  \item{Household overcrowding indicator (rooms)}{Household overcrowding indicator (rooms)}
  \item{Household overcrowding indicator (bedrooms)}{Household overcrowding indicator (bedrooms)}
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd2025_england_lsoa21_indicators
}
\description{
A dataset containing Index of Multiple Deprivation indicators for Lower Layer Super Output Areas (LSOAs) in England.
}
\keyword{datasets}
