% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2025_england_lsoa21_subdomains}
\alias{imd2025_england_lsoa21_subdomains}
\title{Index of Multiple Deprivation (2025) Subdomain Scores for LSOAs (2021) in England}
\format{
A data frame of class "tbl" with the following columns:
\describe{
  \item{lsoa21_code}{LSOA 2021 code}
  \item{children_and_young_people_score}{Score for the Children and Young People subdomain}
  \item{children_and_young_people_rank}{Rank for the Children and Young People subdomain}
  \item{children_and_young_people_decile}{Decile for the Children and Young People subdomain}
  \item{adult_skills_score}{Score for the Adult Skills subdomain}
  \item{adult_skills_rank}{Rank for the Adult Skills subdomain}
  \item{adult_skills_decile}{Decile for the Adult Skills subdomain}
  \item{geographical_barriers_score}{Score for the Geographical Barriers subdomain}
  \item{geographical_barriers_rank}{Rank for the Geographical Barriers subdomain}
  \item{geographical_barriers_decile}{Decile for the Geographical Barriers subdomain}
  \item{wider_barriers_score}{Score for the Wider Barriers subdomain}
  \item{wider_barriers_rank}{Rank for the Wider Barriers subdomain}
  \item{wider_barriers_decile}{Decile for the Wider Barriers subdomain}
  \item{indoors_score}{Score for the Indoors Living Environment subdomain}
  \item{indoors_rank}{Rank for the Indoors Living Environment subdomain}
  \item{indoors_decile}{Decile for the Indoors Living Environment subdomain}
  \item{outdoors_score}{Score for the Outdoors Living Environment subdomain}
  \item{outdoors_rank}{Rank for the Outdoors Living Environment subdomain}
  \item{outdoors_decile}{Decile for the Outdoors Living Environment subdomain}
}
}
\source{
\url{https://www.gov.uk/government/statistics/english-indices-of-deprivation-2025}
}
\usage{
imd2025_england_lsoa21_subdomains
}
\description{
A dataset containing subdomain scores, ranks, and deciles for Lower-layer Super Output Areas (LSOAs) in England.
Subdomains include education (children and young people, adult skills), barriers to housing and services (geographical and wider barriers),
and living environment (indoors and outdoors).
}
\keyword{datasets}
