% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd2025_england_ltla24}
\alias{imd2025_england_ltla24}
\title{Index of Multiple Deprivation (2025) for Lower Tier Local Authorities (2024) in England}
\format{
A data frame of class "tbl" with the following columns:
\describe{
  \item{ltla24_code}{LTLA 2024 code}
  \item{ltla24_name}{LTLA 2024 name}
  \item{IMD - Average rank}{Population-weighted average IMD rank}
  \item{IMD - Rank of average rank}{Rank of the average IMD rank}
  \item{IMD - Average score}{Population-weighted average IMD score}
  \item{IMD - Rank of average score}{Rank of the average IMD score}
  \item{IMD - Proportion of LSOAs in most deprived 10percent nationally}{Proportion of LSOAs in the most deprived 10 percent nationally}
  \item{IMD - Rank of proportion of LSOAs in most deprived 10percent nationally}{Rank of proportion of LSOAs in the most deprived 10 percent nationally}
  \item{IMD25 - Extent}{Extent measure for IMD 2025}
  \item{IMD25 - Rank of extent}{Rank of extent measure}
  \item{IMD25 - Local concentration}{Local concentration measure for IMD 2025}
  \item{IMD25 - Rank of local concentration}{Rank of local concentration measure}
  \item{Income - Average rank}{Population-weighted average Income rank}
  \item{Income - Rank of average rank}{Rank of the average Income rank}
  \item{Income - Average score}{Population-weighted average Income score}
  \item{Income - Rank of average score}{Rank of the average Income score}
  \item{Income - Proportion of LSOAs in most deprived 10percent nationally}{Proportion of LSOAs in the most deprived 10 percent nationally for Income}
  \item{Income - Rank of proportion of LSOAs in most deprived 10percent nationally}{Rank of proportion of LSOAs in the most deprived 10 percent nationally for Income}
  \item{Income scale}{Income scale measure}
  \item{Rank of Income Scale}{Rank of Income scale measure}
  \item{Employment - Average rank}{Population-weighted average Employment rank}
  \item{Employment - Rank of average rank}{Rank of the average Employment rank}
  \item{Employment - Average score}{Population-weighted average Employment score}
  \item{Employment - Rank of average score}{Rank of the average Employment score}
  \item{Employment - Proportion of LSOAs in most deprived 10percent nationally}{Proportion of LSOAs in the most deprived 10 percent nationally for Employment}
  \item{Employment - Rank of proportion of LSOAs in most deprived 10percent nationally}{Rank of proportion of LSOAs in the most deprived 10 percent nationally for Employment}
  \item{Employment scale}{Employment scale measure}
  \item{Rank of Employment Scale}{Rank of Employment scale measure}
  \item{Education, Skills and Training - Average rank}{Population-weighted average Education rank}
  \item{Education, Skills and Training - Rank of average rank}{Rank of the average Education rank}
  \item{Education, Skills and Training - Average score}{Population-weighted average Education score}
  \item{Education, Skills and Training - Rank of average score}{Rank of the average Education score}
  \item{Education, Skills and Training - Proportion of LSOAs in most deprived 10percent nationally}{Proportion of LSOAs in the most deprived 10 percent nationally for Education}
  \item{Education, Skills and Training - Rank of proportion of LSOAs in most deprived 10percent nationally}{Rank of proportion of LSOAs in the most deprived 10 percent nationally for Education}
  \item{Health Deprivation and Disability - Average rank}{Population-weighted average Health rank}
  \item{Health Deprivation and Disability - Rank of average rank}{Rank of the average Health rank}
  \item{Health Deprivation and Disability - Average score}{Population-weighted average Health score}
  \item{Health Deprivation and Disability - Rank of average score}{Rank of the average Health score}
  \item{Health Deprivation and Disability - Proportion of LSOAs in most deprived 10percent nationally}{Proportion of LSOAs in the most deprived 10 percent nationally for Health}
  \item{Health Deprivation and Disability - Rank of proportion of LSOAs in most deprived 10percent nationally}{Rank of proportion of LSOAs in the most deprived 10 percent nationally for Health}
  \item{Crime - Average rank}{Population-weighted average Crime rank}
  \item{Crime - Rank of average rank}{Rank of the average Crime rank}
  \item{Crime - Average score}{Population-weighted average Crime score}
  \item{Crime - Rank of average score}{Rank of the average Crime score}
  \item{Crime - Proportion of LSOAs in most deprived 10percent nationally}{Proportion of LSOAs in the most deprived 10 percent nationally for Crime}
  \item{Crime - Rank of proportion of LSOAs in most deprived 10percent nationally}{Rank of proportion of LSOAs in the most deprived 10 percent nationally for Crime}
  \item{Barriers to Housing and Services - Average rank}{Population-weighted average Barriers rank}
  \item{Barriers to Housing and Services - Rank of average rank}{Rank of the average Barriers rank}
  \item{Barriers to Housing and Services - Average score}{Population-weighted average Barriers score}
  \item{Barriers to Housing and Services - Rank of average score}{Rank of the average Barriers score}
  \item{Barriers to Housing and Services - Proportion of LSOAs in most deprived 10percent nationally}{Proportion of LSOAs in the most deprived 10 percent nationally for Barriers}
  \item{Barriers to Housing and Services - Rank of proportion of LSOAs in most deprived 10percent nationally}{Rank of proportion of LSOAs in the most deprived 10 percent nationally for Barriers}
  \item{Living Environment - Average rank}{Population-weighted average Living Environment rank}
  \item{Living Environment - Rank of average rank}{Rank of the average Living Environment rank}
  \item{Living Environment - Average score}{Population-weighted average Living Environment score}
  \item{Living Environment - Rank of average score}{Rank of the average Living Environment score}
  \item{Living Environment - Proportion of LSOAs in most deprived 10percent nationally}{Proportion of LSOAs in the most deprived 10 percent nationally for Living Environment}
  \item{Living Environment - Rank of proportion of LSOAs in most deprived 10percent nationally}{Rank of proportion of LSOAs in the most deprived 10 percent nationally for Living Environment}
}
}
\source{
\url{https://www.gov.uk/government/statistics/english-indices-of-deprivation-2025}
}
\usage{
imd2025_england_ltla24
}
\description{
A dataset containing IMD scores, ranks, and subdomain measures for Lower Tier Local Authorities (LTLAs) in England.
}
\keyword{datasets}
