% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd_england_msoa}
\alias{imd_england_msoa}
\title{Index of Multiple Deprivation for MSOAs in England}
\format{
A data frame of class "tbl" with 6,791 rows and 25 variables:
\describe{
  \item{msoa_code}{MSOA code}
  \item{Score}{Population-weighted average IMD score}
  \item{Proportion}{Proportion of small areas in the nation's most deprived}
  \item{Extent}{Extent for overall IMD}
  \item{Income_Score}{Population-weighted average score for Income domain}
  \item{Income_Proportion}{Proportion of most deprived areas for Income}
  \item{Income_Extent}{Extent for Income}

  \item{Employment_Score}{Population-weighted average score for Employment
    domain}
  \item{Employment_Proportion}{Proportion of most deprived areas for
    Employment}
  \item{Employment_Extent}{Extent for Employment}

  \item{Education_Score}{Population-weighted average score for Education
    domain}
  \item{Education_Proportion}{Proportion of most deprived areas for
    Education}
  \item{Education_Extent}{Extent for Education}

  \item{Health_Score}{Population-weighted average score for Health domain}
  \item{Health_Proportion}{Proportion of most deprived areas for Health}
  \item{Health_Extent}{Extent for Health}

  \item{Crime_Score}{Population-weighted average score for Crime domain}
  \item{Crime_Proportion}{Proportion of most deprived areas for Crime}
  \item{Crime_Extent}{Extent for Crime}

  \item{Housing_and_Access_Score}{Population-weighted average score for
    Housing/Access domain}
  \item{Housing_and_Access_Proportion}{Proportion of most deprived areas
    for Housing/Access domain}
  \item{Housing_and_Access_Extent}{Extent for Housing/Access}

  \item{Environment_Score}{Population-weighted average score for Environment
    domain}
  \item{Environment_Proportion}{Proportion of most deprived areas for
    Environment}
  \item{Environment_Extent}{Extent for Environment}
  ...
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd_england_msoa
}
\description{
A data set containing IMD ranks and deciles for Middle Layer Super
Output Areas (LSOAs) in England.
}
\keyword{datasets}
