% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd_england_msoa_subdomains}
\alias{imd_england_msoa_subdomains}
\title{Sub-domains of deprivation for MSOAs in England}
\format{
A data frame of class "tbl" with 6,791 rows and 13 variables:
\describe{
  \item{msoa_code}{MSOA code}
  \item{Children_and_young_people_Proportion}{Children and Young People
  sub-domain proportion}
  \item{Children_and_young_people_Extent}{Children and Young People
  sub-domain extent}
  \item{Adult_skills_Proportion}{Adult Skills sub-domain proportion}
  \item{Adult_skills_Extent}{Adult Skills sub-domain extent}
  \item{Geographical_barriers_Proportion}{Geographical Barriers sub-domain
  proportion}
  \item{Geographical_barriers_Extent}{Geographical Barriers sub-domain
  extent}
  \item{Wider_barriers_Proportion}{Wider Barriers sub-domain proportion}
  \item{Wider_barriers_Extent}{Wider Barriers sub-domain extent}
  \item{Indoors_Proportion}{Indoors sub-domain proportion}
  \item{Indoors_Extent}{Indoors sub-domain extent}
  \item{Outdoors_Proportion}{Outdoors sub-domain proportion}
  \item{Outdoors_Extent}{Outdoors sub-domain extent}
  ...
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd_england_msoa_subdomains
}
\description{
A data set containing deprivation sub-domain ranks and deciles for Middle
Layer Super Output Areas (MSOAs) in England.
}
\keyword{datasets}
