% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd_northern_ireland_lad}
\alias{imd_northern_ireland_lad}
\title{Index of Multiple Deprivation for Local Government Districts in Northern
Ireland}
\format{
A data frame of class "tbl" with 11 rows and 25 variables:
\describe{
  \item{lad_code}{LAD code}
  \item{Proportion}{Proportion of small areas in the nation's most deprived}
  \item{Extent}{Extent for overall IMD}
  \item{Income_Proportion}{Proportion of most deprived areas for Income}
  \item{Income_Extent}{Extent for Income}
  \item{Employment_Proportion}{Proportion of most deprived areas for
    Employment}
  \item{Employment_Extent}{Extent for Employment}
  \item{Education_Proportion}{Proportion of most deprived areas for
    Education}
  \item{Education_Extent}{Extent for Education}
  \item{Health_Proportion}{Proportion of most deprived areas for Health}
  \item{Health_Extent}{Extent for Health}
  \item{Crime_Proportion}{Proportion of most deprived areas for Crime}
  \item{Crime_Extent}{Extent for Crime}
  \item{Access_Proportion}{Proportion of most deprived areas
    for Access domain}
  \item{Access_Extent}{Extent for Access}
  \item{Environment_Proportion}{Proportion of most deprived areas for
    Environment}
  \item{Environment_Extent}{Extent for Environment}
  ...
}
}
\source{
\url{https://www.nisra.gov.uk/}
}
\usage{
imd_northern_ireland_lad
}
\description{
A data set containing proportions and extents for Local Government Districts
(Local Authorities) in Northern Ireland.
}
\keyword{datasets}
