% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd_scotland_lad}
\alias{imd_scotland_lad}
\title{Index of Multiple Deprivation for Council Areas in Scotland}
\format{
A data frame of class "tbl" with 32 rows and 17 variables:
\describe{
  \item{lad_code}{Local Authority code}
  \item{Proportion}{Proportion of small areas in the nation's most deprived}
  \item{Extent}{Extent for overall IMD}
  \item{Income_Proportion}{Proportion of most deprived areas for Income}
  \item{Income_Extent}{Extent for Income}
  \item{Employment_Proportion}{Proportion of most deprived areas for
    Employment}
  \item{Employment_Extent}{Extent for Employment}
  \item{Education_Proportion}{Proportion of most deprived areas for
    Education}
  \item{Education_Extent}{Extent for Education}
  \item{Health_Proportion}{Proportion of most deprived areas for Health}
  \item{Health_Extent}{Extent for Health}
  \item{Crime_Proportion}{Proportion of most deprived areas for Crime}
  \item{Crime_Extent}{Extent for Crime}
  \item{Housing_Proportion}{Proportion of most deprived areas
    for Housing domain}
  \item{Housing_Extent}{Extent for Housing}
  \item{Access_Proportion}{Proportion of most deprived areas
    for Access domain}
  \item{Access_Extent}{Extent for Access}
  ...
}
}
\source{
\url{https://www.gov.scot/}
}
\usage{
imd_scotland_lad
}
\description{
A data set containing IMD ranks and deciles for Council Areas (Local
Authorities) in Scotland. Note that IMD scores for Scotland are not
available.
}
\keyword{datasets}
