% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd_uk_lad}
\alias{imd_uk_lad}
\title{Index of Multiple Deprivation (multiple years) for LAD in UK}
\format{
A data frame of class "tbl" with 382 rows and 2 variables:
\describe{
  \item{lad_code}{LAD equivalent code}
  \item{`IMD Deciles, 1: most deprived, 10: least deprived`}{IMD decile}
  ...
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd_uk_lad
}
\description{
A data set containing IMD deciles for the UK. Each nation deciled separately.
}
\keyword{datasets}
