% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd_uk_lsoa}
\alias{imd_uk_lsoa}
\title{Index of Multiple Deprivation (multiple years) for LSOAs across the UK}
\format{
A data frame of class "tbl":
\describe{
  \item{area_code}{LSOA code or devolved equivalent}
  \item{IMD_decile}{IMD decile}
  ...
}
}
\source{
\url{https://www.gov.uk/}
}
\usage{
imd_uk_lsoa
}
\description{
A data set containing IMD deciles for the UK. Each nation deciled separately.
}
\keyword{datasets}
