% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imd_wales_lsoa}
\alias{imd_wales_lsoa}
\title{Index of Multiple Deprivation for LSOAs in Wales}
\format{
A data frame of class "tbl" with 1,909 rows and 19 variables:
\describe{
  \item{lsoa_code}{LSOA code}
  \item{IMD_decile}{IMD decile}
  \item{Income_decile}{Income domain decile}
  \item{Employment_decile}{Employment domain decile}
  \item{Education_decile}{Education domain decile}
  \item{Health_decile}{Health and Disability domain decile}
  \item{Crime_decile}{Crime domain decile}
  \item{Housing_decile}{Housing domain decile}
  \item{Access_decile}{Access to services domain decile}
  \item{Environment_decile}{Environment domain decile}
  \item{IMD_rank}{IMD rank}
  \item{Income_rank}{Income domain rank}
  \item{Employment_rank}{Employment domain rank}
  \item{Education_rank}{Education domain rank}
  \item{Health_rank}{Health domain rank}
  \item{Crime_rank}{Crime domain rank}
  \item{Housing_rank}{Housing domain rank}
  \item{Access_rank}{Access to services domain rank}
  \item{Environment_rank}{Environment domain rank}
  ...
}
}
\source{
\url{https://gov.wales/}
}
\usage{
imd_wales_lsoa
}
\description{
A data set containing IMD ranks and deciles for Lower Layer Super
Output Areas (LSOAs) in Wales.
}
\keyword{datasets}
