% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.r
\name{invert_this}
\alias{invert_this}
\title{Invert deciles, ranks, percentiles etc.
e.g. with decile, a score of 10 --> 1 and score of 1 --> 10}
\usage{
invert_this(x)
}
\arguments{
\item{x}{Vector of data to invert}
}
\description{
Invert deciles, ranks, percentiles etc.
e.g. with decile, a score of 10 --> 1 and score of 1 --> 10
}
