% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_composite_imd.R
\name{load_composite_imd}
\alias{load_composite_imd}
\title{Load composite 2020 UK Index of Multiple Deprivation}
\usage{
load_composite_imd(nation = "E")
}
\arguments{
\item{nation}{The initial letter of the UK nation whose composite IMD you
want to load. Should be one of: (E)ngland, (W)ales, (S)cotland
(N)orthern Ireland. See the GitHub link about for more details
about why there are separate composite IMDs for each nation.}
}
\description{
Load data from mySociety's composite 2020 UK Index of Multiple Deprivation:
https://github.com/mysociety/composite_uk_imd.
}
\examples{
\dontrun{
imd_uk <- load_composite_imd("E")
}
}
