% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_weighted_scores.R
\name{pop_weighted_scores}
\alias{pop_weighted_scores}
\title{Calculate population-weighted scores and extent}
\usage{
pop_weighted_scores(
  d,
  domain = "Index of Multiple Deprivation (IMD)",
  score_suffix = " Score",
  rank_suffix = " Rank (where 1 is most deprived)",
  decile_suffix = " Decile (where 1 is most deprived 10\% of LSOAs)",
  higher_level_geography,
  population_col = "No. people"
)
}
\arguments{
\item{d}{Dataframe containing ranks, scores and population estimates}

\item{domain}{Which domain to calculate population-weighted scores for}

\item{score_suffix}{The wording of the variable containing scores}

\item{rank_suffix}{The wording of the variable containing ranks}

\item{decile_suffix}{The wording of the variable containing deciles}

\item{higher_level_geography}{Name of the variable containing the
higher-level geography names/codes}

\item{population_col}{Name of the column containing population estimates}
}
\description{
Calculate the Extent and population-weighted average scores for
    higher-level geographies

"Extent" is the proportion of the local population that live in areas
classified as among the most deprived in the higher geography. To calculate
this, we need to first calculate a weighted score based on the population in
the most deprived 30% of areas
}
