% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_weighted_scores.R
\name{calculate_pop_weighted_score}
\alias{calculate_pop_weighted_score}
\title{Calculate population-weighted scores and extent}
\usage{
calculate_pop_weighted_score(data, var, higher_level_geography, population)
}
\arguments{
\item{data}{Data frame containing a variable to be aggregated, lower level
geography population estimates, and a higher level geographical
grouping variable}

\item{var}{Name of the variable in the data frame containing the variable to
be aggregated (e.g. rank) for the lower level geography}

\item{higher_level_geography}{Name of the variable in the data frame
containing the higher level geography names/codes}

\item{population}{Name of the variable in the data frame containing
the population estimates of the lower level geography}

\item{d}{Dataframe containing ranks, scores and population estimates}

\item{domain}{Which domain to calculate population-weighted scores for}

\item{score_suffix}{The wording of the variable containing scores}

\item{rank_suffix}{The wording of the variable containing ranks}

\item{decile_suffix}{The wording of the variable containing deciles}

\item{population_col}{Name of the column containing population estimates

Population-weighted scores

Calculate population-weighted scores within small areas.}
}
\description{
Calculate the Extent and population-weighted average scores for
    higher-level geographies

"Extent" is the proportion of the local population that live in areas
classified as among the most deprived in the higher geography. To calculate
this, we need to first calculate a weighted score based on the population in
the most deprived 30% of areas
}
\examples{
\dontrun{
calculate_pop_weighted_score(imd_england_lsoa, IMD_score, pop_count)
}
}
